Clean: up merge.py file (#705)
Closes: https://github.com/Granulate/gprofiler/issues/672
