+++ tests/test_node.py
@@ -9,9 +9,9 @@ import pytest
 from docker import DockerClient
 from docker.models.images import Image
 
-from gprofiler.merge import parse_one_collapsed
 from gprofiler.profiler_state import ProfilerState
 from gprofiler.profilers.perf import SystemProfiler
+from gprofiler.utils.collapsed_format import parse_one_collapsed
 from tests import CONTAINERS_DIRECTORY
 from tests.conftest import AssertInCollapsed
 from tests.utils import (
